+++ gance/overlay/overlay_common.py
@@ -1,0 +1,159 @@
+"""
+Write parts of one video on top of another.
+Do things like track eyes to interesting effect.
+"""
+
+import itertools
+from typing import Dict, Iterator, List, NamedTuple, Optional, Tuple, cast
+
+import numpy as np
+from cv2 import cv2
+from PIL import Image, ImageDraw
+from scipy.spatial import distance
+
+from gance.gance_types import RGBInt8ImageType
+from gance.image_sources.image_sources_common import ImageResolution, image_resolution
+from gance.overlay.overlay_visualization import OverlayContext
+
+
+class BoundingBox(NamedTuple):
+    """
+    Describes a bounding box rectangle as output by opencv.
+    """
+
+    x: int
+    y: int
+    width: int
+    height: int
+
+
+def landmarks_to_bounding_boxes(landmarks: List[Dict[str, Tuple[int, ...]]]) -> List[BoundingBox]:
+    """
+    For each of the sets of keypoints, pull out the interesting ones, and draw a bounding box
+    around them. Currently, the left and right eye keypoints are used.
+    :param landmarks: From the face recognition library.
+    :return: List of bounding boxes.
+    """
+
+    return [
+        BoundingBox(*cv2.boundingRect(np.array(landmark["left_eye"] + landmark["right_eye"])))
+        for landmark in landmarks
+    ]
+
+
+def bounding_box_center(bounding_box: BoundingBox) -> Tuple[float, float]:
+    """
+    Finds the center x,y coordinate of a given bounding box.
+    :param bounding_box: Box to analyze.
+    :return: (x,y) of center.
+    """
+
+    return (bounding_box.x + bounding_box.width / 2), (bounding_box.y + bounding_box.height / 2)
+
+
+def bounding_box_distance(
+    a_boxes: List[BoundingBox], b_boxes: List[BoundingBox]
+) -> Optional[float]:
+    """
+    Calculate the minimum distance between two sets of bounding boxes.
+    :param a_boxes: Left side.
+    :param b_boxes: Right side.
+    :return: Minimum distance between the origins of these boxes in pixels.
+    """
+
+    if a_boxes and b_boxes:
+        # Only take the X,Y points here
+        a_origin = [bounding_box_center(box) for box in a_boxes]
+        b_origin = [bounding_box_center(box) for box in b_boxes]
+
+        return min(
+            [
+                float(distance.euclidean(target_point, final_point))
+                for target_point, final_point in itertools.product(a_origin, b_origin)
+            ]
+        )
+    else:
+        return None
+
+
+def _draw_mask(resolution: ImageResolution, bounding_boxes: List[BoundingBox]) -> "Image":
+    """
+    Draw bounding boxes as a white mask. Edges of bounding boxes are
+    included in mask.
+    :param resolution: The size of the resulting image.
+    :param bounding_boxes: To draw as a mask on the image.
+    :return: The input image, but now with the bounding boxes down onto it.
+    """
+
+    output = Image.new("L", tuple(resolution))
+    draw = ImageDraw.Draw(output)
+
+    for bounding_box in bounding_boxes:
+
+        x, y, w, h = bounding_box
+
+        y_center = y + (h / 2)
+        y_lower = y_center + 60
+        y_upper = y_center - 60
+        x_left = x - 200
+        x_right = x + (w + 200)
+
+        draw.polygon(
+            [(x_left, y_lower), (x_right, y_lower), (x_right, y_upper), (x_left, y_upper)],
+            outline=255,
+            fill=255,
+        )
+
+    return output
+
+
+def _apply_mask(
+    foreground_image: "Image", background_image: RGBInt8ImageType, mask: "Image"
+) -> RGBInt8ImageType:
+    """
+    Writes a masked region of the foreground onto the background.
+    :param foreground_image: An image that has already been converted to a PIL Image. This will
+    be drawn onto the background.
+    :param background_image: The region of `foreground_image` will be drawn onto this image.
+    :param mask: Describes the region of `foreground_image` to draw onto `background_image`.
+    :return: The new image in the canonical type.
+    """
+    return cast(
+        RGBInt8ImageType,
+        np.asarray(Image.composite(foreground_image, Image.fromarray(background_image), mask)),
+    )
+
+
+def write_boxes_onto_image(
+    foreground_image: RGBInt8ImageType,
+    background_image: RGBInt8ImageType,
+    bounding_boxes: List[BoundingBox],
+) -> RGBInt8ImageType:
+    """
+    Write regions of the foreground image onto the background image.
+    Return the result as a new image.
+    :param foreground_image: To be drawn onto the background.
+    :param background_image: The be drawn under the foreground.
+    :param bounding_boxes: The regions of the foreground to be drawn onto the background.
+    :return: A new image, with the foreground regions written onto the background.
+    """
+
+    foreground = Image.fromarray(foreground_image)
+
+    return _apply_mask(
+        foreground_image=foreground,
+        background_image=background_image,
+        mask=_draw_mask(image_resolution(foreground_image), bounding_boxes),
+    )
+
+
+class OverlayResult(NamedTuple):
+    """
+    The different output streams from an eye tracking overlay computation.
+    See the docs for `_FrameOverlayResult` for meaning as to what the different
+    members are here, these are iterators of those types.
+    Note: really important that the order of the members matches `_FrameOverlayResult`.
+    """
+
+    bbox_lists: Iterator[Optional[List[BoundingBox]]]
+    contexts: Iterator[OverlayContext]
