+++ setup.py
@@ -16,7 +16,7 @@ def prod_dependencies() -> List[str]:
 
 setup(
     name="gance",
-    version="0.16.1",
+    version="0.18.0",
     description="Maps music and video into the latent space of StyleGAN models.",
     author="Devon Bray",
     author_email="dev@esologic.com",
