+++ test/test_overlay.py
@@ -4,8 +4,8 @@ Happy path tests
 
 from test.assets import BATCH_2_IMAGE_1_PATH, BATCH_2_IMAGE_2_PATH
 
-from gance import overlay
 from gance.image_sources import still_image_common
+from gance.overlay import overlay_common
 
 
 def test_write_boxes_onto_image() -> None:
@@ -14,8 +14,8 @@ def test_write_boxes_onto_image() -> None:
     :return: None
     """
 
-    overlay.write_boxes_onto_image(
+    overlay_common.write_boxes_onto_image(
         foreground_image=still_image_common.read_image(image_path=BATCH_2_IMAGE_1_PATH),
         background_image=still_image_common.read_image(image_path=BATCH_2_IMAGE_2_PATH),
-        bounding_boxes=[overlay.BoundingBox(x=0, y=0, width=100, height=100)],
+        bounding_boxes=[overlay_common.BoundingBox(x=0, y=0, width=100, height=100)],
     )
